requirements.txt
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.59.2
greenlet==3.2.4
idna==3.10
iniconfig==2.1.0
keyboard==0.13.5
kiwisolver==1.4.9
lxml==6.0.1
matplotlib==3.10.6
MouseInfo==0.1.3
mss==10.1.0
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pillow==10.4.0
playwright==1.55.0
pluggy==1.6.0
PyAutoGUI==0.9.54
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytweening==1.2.0
pytz==2025.2
pywin32==311
PyYAML==6.0.2
requests==2.32.5
six==1.17.0
soupsieve==2.8
ttkbootstrap==1.14.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xlrd==1.2.0
